カスタムフックを使ってリファクタリング
